Conflicts:
	idea/src/org/jetbrains/jet/plugin/completion/weigher/JetCompletionSorting.kt
	idea/src/org/jetbrains/jet/plugin/conversion/copy/ConvertJavaCopyPastePostProcessor.kt
	idea/tests/org/jetbrains/jet/plugin/JetLightCodeInsightFixtureTestCase.java
	update_dependencies.xml
	JavaRuntimeDetectionUtil
        JsLibraryStdDetectionUtil
        CompletionSorting
        ...

+++ js/js.translator/testData/webDemoCanvasExamples/cases/Fancy lines.kt
@@ -43,7 +43,7 @@ class FancyLines() {
         context.bezierCurveTo(width * Math.random(), height * Math.random(),
                               width * Math.random(), height * Math.random(), x, y);
 
-        hue += Math.random() * 10;
+        hue += (Math.random() * 10).toInt();
 
         context.strokeStyle = "hsl($hue, 50%, 50%)";
